Split opaque out from brand
