+++ packages/dds/tree/src/test/util/opaque.spec.ts
@@ -4,20 +4,23 @@
  */
 
 import {
-	Brand,
-	brand,
-	Erased,
-	fromErased,
-	brandErased,
+	Opaque,
+	ExtractFromOpaque,
+	brandOpaque,
+	extractFromOpaque,
 	// Allow importing from this specific file which is being tested:
 	/* eslint-disable-next-line import/no-internal-modules */
-} from "../../util/brand.js";
+} from "../../util/opaque.js";
 
 import {
+	Brand,
+	BrandedType,
+	Erased,
 	areSafelyAssignable,
 	isAssignableTo,
 	requireTrue,
 	requireFalse,
+	isAny,
 } from "../../util/index.js";
 
 // These tests currently just cover the type checking, so its all compile time.
@@ -25,35 +28,26 @@ import {
 export type T1 = Brand<number, "1">;
 export type T2 = Brand<number, "2">;
 
-type Intersected = T1 & T2;
-type T1Constant = Brand<4, "1">;
+interface O1 extends Opaque<T1> {}
+interface O2 extends Opaque<T2> {}
 
 type _check =
+	| requireTrue<areSafelyAssignable<ExtractFromOpaque<O1>, T1>>
+	| requireTrue<areSafelyAssignable<ExtractFromOpaque<O2>, T2>>
 	// Check covariant ValueType
-	| requireTrue<isAssignableTo<T1Constant, T1>>
-	| requireFalse<isAssignableTo<T1, T2>>
-	// Check multiple brands don't produce never.
-	| requireFalse<isAssignableTo<Intersected, never>>;
+	| requireFalse<isAssignableTo<O1, O2>>;
 
-const _branded: T1 = brand(0);
-
-// @ts-expect-error No type to infer: does not build.
-const _branded2 = brand(0);
+const _opaque: O1 = brandOpaque<O1>(0);
 
 // @ts-expect-error No type to infer: does not build.
 const untypedOpaque = brandOpaque(0);
+
+// If somehow an untyped opaque handle is produced, make sure any does not leak out:
+const extracted = extractFromOpaque(0 as any as BrandedType<any, string>);
+type _check2 = requireFalse<isAny<typeof extracted>>;
 
 // Erased
 interface E4 extends Erased<"4"> {}
 interface E5 extends Erased<"5"> {}
 export type T4 = Brand<{ test: number }, E4>;
 export type T5 = Brand<{ test: number }, E5>;
-
-const erased = brandErased<T4>({ test: 5 });
-const branded = fromErased<T4>(erased);
-type _check1 =
-	| requireTrue<areSafelyAssignable<typeof erased, E4>>
-	| requireTrue<areSafelyAssignable<typeof branded, T4>>
-	// Check strong typing
-	| requireFalse<isAssignableTo<E4, E5>>
-	| requireFalse<isAssignableTo<T4, T5>>;
