+++ packages/dds/tree/src/util/index.ts
@@ -7,19 +7,16 @@ export {
 	brand,
 	Brand,
 	BrandedType,
-	brandOpaque,
 	brandedNumberType,
 	brandedStringType,
-	extractFromOpaque,
-	ExtractFromOpaque,
 	fromErased,
 	Erased,
 	ErasedType,
 	NameFromBranded,
-	Opaque,
 	ValueFromBranded,
 	brandErased,
 } from "./brand.js";
+export { brandOpaque, extractFromOpaque, ExtractFromOpaque, Opaque } from "./opaque.js";
 export {
 	deleteFromNestedMap,
 	getOrAddInMap,
@@ -73,7 +70,6 @@ export {
 	Populated,
 	RecursiveReadonly,
 	zipIterables,
-	Assume,
 	assertValidIndex,
 	assertValidRange,
 	assertNonNegativeSafeInteger,
@@ -100,6 +96,7 @@ export {
 	FlattenKeys,
 	AllowOptionalNotFlattened,
 	RestrictiveReadonlyRecord,
+	Assume,
 } from "./typeUtils.js";
 
 export {
