+++ packages/dds/tree/src/util/utils.ts
@@ -295,20 +295,6 @@ export function assertNonNegativeSafeInteger(index: number) {
 }
 
 /**
- * Assume that `TInput` is a `TAssumeToBe`.
- *
- * @remarks
- * This is useful in generic code when it is impractical (or messy)
- * to to convince the compiler that a generic type `TInput` will extend `TAssumeToBe`.
- * In these cases `TInput` can be replaced with `Assume<TInput, TAssumeToBe>` to allow compilation of the generic code.
- * When the generic code is parameterized with a concrete type, if that type actually does extend `TAssumeToBe`,
- * it will behave like `TInput` was used directly.
- *
- * @internal
- */
-export type Assume<TInput, TAssumeToBe> = [TInput] extends [TAssumeToBe] ? TInput : TAssumeToBe;
-
-/**
  * Convert an object into a Map.
  *
  * This function must only be used with objects specifically intended to encode map like information.
